refactor(config): split db runtime into @alesha-nov/db
Move Bun SQL client, migration runner, and auth migration bundle into a dedicated @alesha-nov/db package so @alesha-nov/config stays focused on environment-driven auth/email config resolution.
