+++ .github/workflows/pr-checks.yml
@@ -65,6 +65,7 @@ jobs:
       - name: Lint all packages
         run: |
           bun run --filter @alesha-nov/config lint && \
+          bun run --filter @alesha-nov/db lint && \
           bun run --filter @alesha-nov/email lint && \
           bun run --filter @alesha-nov/auth lint && \
           bun run --filter @alesha-nov/auth-web lint && \
@@ -93,6 +94,7 @@ jobs:
       - name: Build packages
         run: |
           bun run --filter @alesha-nov/config build && \
+          bun run --filter @alesha-nov/db build && \
           bun run --filter @alesha-nov/email build && \
           bun run --filter @alesha-nov/auth build && \
           bun run --filter @alesha-nov/auth-web build && \
@@ -108,6 +110,7 @@ jobs:
 
           bun test --coverage --coverage-reporter=lcov --reporter=junit --reporter-outfile=./test-results/junit.xml \
             packages/config/src \
+            packages/db/src \
             packages/email/src \
             packages/auth/src \
             packages/auth-web/src
