+++ docs/_coverpage.md
@@ -6,7 +6,8 @@ Bun-powered TypeScript monorepo for authentication, email, and configuration pac
 
 | Package | Description |
 |---------|-------------|
-| [`@alesha-nov/config`](packages/config) | Database client, migrations, env helpers |
+| [`@alesha-nov/config`](packages/config) | Auth/email env config helpers |
+| [`@alesha-nov/db`](packages/db) | Database client and migration helpers |
 | [`@alesha-nov/auth`](packages/auth) | Auth service: signup, login, magic links, OAuth |
 | [`@alesha-nov/email`](packages/email) | Email via AWS SES or SMTP |
 | [`@alesha-nov/auth-web`](packages/auth-web) | HTTP route handlers + cookie sessions |
