+++ package.json
@@ -7,10 +7,10 @@
     "apps/*"
   ],
   "scripts": {
-    "build": "bun run --filter @alesha-nov/config build && bun run --filter @alesha-nov/email build && bun run --filter @alesha-nov/auth build && bun run --filter @alesha-nov/auth-web build && bun run --filter @alesha-nov/auth-react build && bun run --filter web build",
-    "typecheck": "bun run --filter @alesha-nov/config typecheck && bun run --filter @alesha-nov/email typecheck && bun run --filter @alesha-nov/auth typecheck && bun run --filter @alesha-nov/auth-web typecheck && bun run --filter @alesha-nov/auth-react typecheck && bun run --filter web typecheck",
-    "lint": "bun run --filter @alesha-nov/config lint && bun run --filter @alesha-nov/email lint && bun run --filter @alesha-nov/auth lint && bun run --filter @alesha-nov/auth-web lint && bun run --filter @alesha-nov/auth-react lint && bun run --filter web lint",
-    "test": "bun run --filter @alesha-nov/config test && bun run --filter @alesha-nov/email test && bun run --filter @alesha-nov/auth test && bun run --filter @alesha-nov/auth-web test && bun run --filter @alesha-nov/auth-react test && bun run --filter web test",
+    "build": "bun run --filter @alesha-nov/config build && bun run --filter @alesha-nov/db build && bun run --filter @alesha-nov/email build && bun run --filter @alesha-nov/auth build && bun run --filter @alesha-nov/auth-web build && bun run --filter @alesha-nov/auth-react build && bun run --filter web build",
+    "typecheck": "bun run --filter @alesha-nov/config typecheck && bun run --filter @alesha-nov/db typecheck && bun run --filter @alesha-nov/email typecheck && bun run --filter @alesha-nov/auth typecheck && bun run --filter @alesha-nov/auth-web typecheck && bun run --filter @alesha-nov/auth-react typecheck && bun run --filter web typecheck",
+    "lint": "bun run --filter @alesha-nov/config lint && bun run --filter @alesha-nov/db lint && bun run --filter @alesha-nov/email lint && bun run --filter @alesha-nov/auth lint && bun run --filter @alesha-nov/auth-web lint && bun run --filter @alesha-nov/auth-react lint && bun run --filter web lint",
+    "test": "bun run --filter @alesha-nov/config test && bun run --filter @alesha-nov/db test && bun run --filter @alesha-nov/email test && bun run --filter @alesha-nov/auth test && bun run --filter @alesha-nov/auth-web test && bun run --filter @alesha-nov/auth-react test && bun run --filter web test",
     "test:coverage": "bun test --coverage --coverage-reporter=lcov",
     "changeset": "changeset",
     "version-packages": "changeset version",
