+++ packages/auth/package.json
@@ -41,7 +41,7 @@
     "test": "bun test"
   },
   "dependencies": {
-    "@alesha-nov/config": "workspace:*",
+    "@alesha-nov/db": "workspace:*",
     "@alesha-nov/email": "workspace:*"
   },
   "repository": {
