+++ packages/auth/src/service.oauth-pkce.test.ts
@@ -11,7 +11,7 @@ const sqlTag = ((strings: TemplateStringsArray, ...values: unknown[]) => {
   unsafe?: (query: string) => Promise<unknown[]>;
 };
 
-mock.module("@alesha-nov/config", () => ({
+mock.module("@alesha-nov/db", () => ({
   createDatabaseClient: () => ({ sql: sqlTag }),
   runMigrations: async () => {},
 }));
