+++ packages/config/README.md
@@ -1,27 +1,24 @@
 # @alesha-nov/config
 
-Database primitives and auth/email configuration resolvers for the monorepo.
+Auth and email environment configuration resolvers for the monorepo.
 
 ## Implemented Features
 
-- `resolveDBType()` for `mysql | postgresql | sqlite`
-- `createDatabaseClient(DBConfig)` using Bun SQL
-- `ensureMigrationsTable()` for `alesha_migrations`
-- `runMigrations()` idempotent migration execution
-- Exported types: `DBType`, `DBConfig`, `Migration`, `DatabaseClient`
 - Auth config types (`JWTConfig`, `SessionConfig`)
 - `resolveJWTSecret()`
 - `resolveSessionConfig()`
 - `resolveMagicLinkConfig()`
 - `resolveEmailTransportConfig()` (SES/SMTP with validation)
 - `resolveOAuthConfig()` (Google/GitHub)
-- `authMigrationsBundle` export for auth package consumption
 
 ## Required for Target Auth (Email/Password, Magic Link, Google/GitHub)
 
-- Stable DB connection for all auth packages
-- Shared migration runner used by `@alesha-nov/auth`
-- Environment-driven DB + auth/email transport resolution
+- Environment-driven auth/email transport resolution
+
+## Scope
+
+- This package is runtime-agnostic and only resolves config values.
+- Database client and migration primitives live in `@alesha-nov/db`.
 
 ## Remaining / Follow-up
