+++ packages/db/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@alesha-nov/config",
-  "version": "0.2.1",
+  "name": "@alesha-nov/db",
+  "version": "0.1.0",
   "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -8,13 +8,17 @@
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
+    },
+    "./auth-migrations": {
+      "types": "./dist/auth-migrations.d.ts",
+      "import": "./dist/auth-migrations.js"
     }
   },
   "files": [
     "dist"
   ],
   "scripts": {
-    "build": "tsup src/index.ts --format esm --dts",
+    "build": "tsup src/index.ts src/auth-migrations.ts --format esm --dts --external bun",
     "typecheck": "tsc -p tsconfig.json --noEmit",
     "lint": "eslint src --ext .ts",
     "test": "bun test"
@@ -22,23 +26,25 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/JuniYadi/alesha-nov.git",
-    "directory": "packages/config"
+    "directory": "packages/db"
   },
   "license": "MIT",
   "author": "JuniYadi",
   "keywords": [
     "alesha",
-    "config",
-    "env",
-    "configuration",
-    "auth"
+    "db",
+    "database",
+    "bun",
+    "sql",
+    "migrations"
   ],
   "homepage": "https://github.com/JuniYadi/alesha-nov#readme",
   "bugs": {
     "url": "https://github.com/JuniYadi/alesha-nov/issues"
   },
   "engines": {
-    "node": ">=22.12.0"
+    "node": ">=22.12.0",
+    "bun": ">=1.3.0"
   },
   "publishConfig": {
     "access": "public"
